Implemented PID Control
Implemented a PID class. Running script as is will make the turtlebot move in a square. However, issues with turtlebot odometry (and possibly the wheels having uneven torques?) are causing errors in rotation and movement; i.e. the script will move the turtlebot in a square but it will not end exactly where it started off. This is because of small rotational errors that accumulate as the turtlebot continually moves and rotates. Will not be able to fix this until we have access to a more accurate positioning system