Initial CLBF interface commit
Starting to work on functionality to implement CLBF controller on the turtlebots. Very much a work in progress; will be the bridge between the turtlebot hardware scripts and the real_time_series_experiment.py file in the neural_clbf repository. This script was based on the eval_linear_satellite.py script. A few small modifications and TODOs have been made thus far.